+++ features/support/env.rb
@@ -17,7 +17,7 @@ Conjur.configuration.account = ENV['CONJUR_ACCOUNT'] || 'cucumber'
 Conjur.configuration.authn_local_socket = "/run/authn-local/.socket"
 
 $username = ENV['CONJUR_AUTHN_LOGIN'] || 'admin'
-$password = ENV['CONJUR_AUTHN_API_KEY'] || 'secret'
-
-$api_key = Conjur::API.login $username, $password
+$api_key = (ENV['CONJUR_AUTHN_API_KEY'] && !ENV['CONJUR_AUTHN_API_KEY'].empty?) \
+           ? ENV['CONJUR_AUTHN_API_KEY'] \
+           : Conjur::API.login($username, 'secret')
 $conjur = Conjur::API.new_from_key $username, $api_key